+++ tests/core/tests/test_mixins.py
@@ -182,6 +182,62 @@ class MixinModelAdminTest(TestCase):
         admin.get_export_resource_kwargs(self.request)
         self.assertEqual(1, admin.call_count)
 
+    class BaseModelResourceClassTest(mixins.BaseImportMixin, mixins.BaseExportMixin):
+        resource_class = resources.Resource
+        export_call_count = 0
+        import_call_count = 0
+
+        def get_export_resource_class(self):
+            self.export_call_count += 1
+
+        def get_import_resource_class(self):
+            self.import_call_count += 1
+
+    def test_deprecated_resource_class_raises_warning(self):
+        """Test that the mixin throws error if user didn't
+        migrate to resource_classes"""
+        admin = self.BaseModelResourceClassTest()
+        with self.assertWarnsRegex(
+            DeprecationWarning,
+            r"^The 'get_export_resource_class\(\)' method has been deprecated. "
+            r"Please implement the new 'get_export_resource_classes\(\)' method$",
+        ):
+            admin.get_export_resource_classes(self.request)
+
+        with self.assertWarnsRegex(
+            DeprecationWarning,
+            r"^The 'get_import_resource_class\(\)' method has been deprecated. "
+            r"Please implement the new 'get_import_resource_classes\(\)' method$",
+        ):
+            admin.get_import_resource_classes(self.request)
+
+        with self.assertWarnsRegex(
+            DeprecationWarning,
+            r"^The 'resource_class' field has been deprecated. "
+            r"Please implement the new 'resource_classes' field$",
+        ):
+            self.assertEqual(
+                admin.get_resource_classes(self.request), [resources.Resource]
+            )
+
+        self.assertEqual(1, admin.export_call_count)
+        self.assertEqual(1, admin.import_call_count)
+
+    class BaseModelGetExportResourceClassTest(mixins.BaseExportMixin):
+        def get_resource_class(self):
+            pass
+
+    def test_deprecated_get_resource_class_raises_warning(self):
+        """Test that the mixin throws error if user
+        didn't migrate to resource_classes"""
+        admin = self.BaseModelGetExportResourceClassTest()
+        with self.assertWarnsRegex(
+            DeprecationWarning,
+            r"^The 'get_resource_class\(\)' method has been deprecated. "
+            r"Please implement the new 'get_resource_classes\(\)' method$",
+        ):
+            admin.get_resource_classes(self.request)
+
     class BaseModelAdminFaultyResourceClassesTest(mixins.BaseExportMixin):
         resource_classes = resources.Resource
 
@@ -198,6 +254,14 @@ class MixinModelAdminTest(TestCase):
 
         resource_class = resources.Resource
         resource_classes = [resources.Resource]
+
+    def test_both_resource_class_raises_exception(self):
+        """Test fallback mechanism to old get_export_resource_class() method"""
+        admin = self.BaseModelAdminBothResourceTest()
+        with self.assertRaisesRegex(
+            Exception, "Only one of 'resource_class' and 'resource_classes' can be set"
+        ):
+            admin.get_export_resource_classes(self.request)
 
     class BaseModelExportChooseTest(mixins.BaseExportMixin):
         resource_classes = [resources.Resource, FooResource]
@@ -230,6 +294,34 @@ class MixinModelAdminTest(TestCase):
 
         form.cleaned_data = {"resource": 1}
         self.assertEqual(admin.choose_import_resource_class(form, request), FooResource)
+
+    class BaseModelResourceClassOldTest(mixins.BaseImportMixin, mixins.BaseExportMixin):
+        def get_resource_class(self):
+            return FooResource
+
+    def test_get_resource_class_old(self):
+        """
+        Test that if only the old get_resource_class() method is defined,
+        the get_export_resource_classes() and get_import_resource_classes()
+        still return list of resources.
+        """
+        admin = self.BaseModelResourceClassOldTest()
+        with self.assertWarnsRegex(
+            DeprecationWarning,
+            r"^The 'get_resource_class\(\)' method has been deprecated. "
+            r"Please implement the new 'get_resource_classes\(\)' method$",
+        ):
+            self.assertEqual(
+                admin.get_export_resource_classes(self.request), [FooResource]
+            )
+        with self.assertWarnsRegex(
+            DeprecationWarning,
+            r"^The 'get_resource_class\(\)' method has been deprecated. "
+            r"Please implement the new 'get_resource_classes\(\)' method$",
+        ):
+            self.assertEqual(
+                admin.get_import_resource_classes(self.request), [FooResource]
+            )
 
 
 class BaseExportMixinTest(TestCase):